Split IocDataAccessModule.cs functionality into smaller, more specific classes.
