+++ src/VolleyManagement.Backend/VolleyManagement.Data.MsSql/VolleyManagement.Data.MsSql.csproj
@@ -221,6 +221,9 @@
     <Compile Include="Entities\ValidationConstants.cs" />
     <Compile Include="Entities\ContributorEntity.cs" />
     <Compile Include="Infrastructure\IocDataAccessModule.cs" />
+    <Compile Include="Infrastructure\QueriesRegistrator.cs" />
+    <Compile Include="Infrastructure\RepositoriesRegistrator.cs" />
+    <Compile Include="Infrastructure\UserRelatedRepositoriesRegistrator.cs" />
     <Compile Include="Mappers\DomainToDal.cs" />
     <Compile Include="Entities\PlayerEntity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
